feat(did): Add `canisters` package declarations to `did` script (#1294)
# Motivation

We want to start migrate the `canisters` sub-packages into the main
folder. However, first, we need to include the package among the ones
that are run in the declarations script.

We use the same patterns developed by @peterpeterparker in
[Juno](https://github.com/junobuild/juno-js/tree/main/packages/functions/src/canisters)
and [OISY
signer](https://github.com/dfinity/oisy-wallet-signer/tree/main/src/declarations).

# Changes

- Adapt the script to accept a sub-folde declarations (instead of
candid, that still will stay the default until the end of the
migration).
- Add the `canisters` package in the script.

# Tests

Script `npm run did` worked as usual.