feat(canisters): Export declarations and factory of `assets` canister (#1299)
# Motivation

Copying what @peterpeterparker 's did in
[Juno](https://github.com/junobuild/juno-js/tree/main/packages/functions/src/canisters/declarations)
and [OISY
signer](https://github.com/dfinity/oisy-wallet-signer/tree/main/src/declarations),
we create the root modules to export the declarations and the factory
for the `assets` canister.

This type of structure provides us with a more modular and more direct
usage of the bindings.